requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Markdown==0.3
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.0
psycopg2==2.9.2
PyJWT==1.6.4
requests==2.25.1
urllib3==1.26.5
Werkzeug==0.15.5
